finished refactoring account
