+++ src/components/ApplicationViews.js
@@ -7,6 +7,7 @@ import useSimpleAuth from "../hooks/ui/useSimpleAuth"
 import ProductSellForm from "./Product/ProductSellForm"
 import MyAccount from "./Payment/MyAccount"
 import PaymentForm from "./Payment/PaymentForm"
+import PaymentTypeList from "./Payment/PaymentTypeList"
 
 const ApplicationViews = () => {
     return (
@@ -28,12 +29,16 @@ const ApplicationViews = () => {
                 return <ProductSellForm {...props} />
             }}
             />
-            <Route exact path="/payment_types" render={props => {
+            <Route exact path="/my_account" render={props => {
                 return <MyAccount {...props} />
             }}
             />
             <Route exact path="/payment_types/new" render={props => {
                 return <PaymentForm {...props} />
+            }}
+            />
+            <Route exact path="/payment_types" render={props => {
+                return <PaymentTypeList {...props} />
             }}
             />
         </React.Fragment>
